more namespace refactor and cleanup
