+++ Core/Objects/Airport/AirportObject.cs
@@ -3,21 +3,82 @@ using OpenLoco.Dat.Data;
 using OpenLoco.Dat.FileParsing;
 using OpenLoco.Dat.Types;
 
-namespace Core.Objects
+namespace OpenLoco.Dat.Objects
 {
-	public enum AirportMovementNodeFlags : uint16_t
-	{
-		None = 0,
-		Terminal = 1 << 0,
-		TakeoffEnd = 1 << 1,
-		Flag2 = 1 << 2,
-		Taxiing = 1 << 3,
-		InFlight = 1 << 4,
-		HeliTakeoffBegin = 1 << 5,
-		TakeoffBegin = 1 << 6,
-		HeliTakeoffEnd = 1 << 7,
-		Touchdown = 1 << 8,
-	};
+	//[TypeConverter(typeof(ExpandableObjectConverter))]
+	//[LocoStructType(ObjectType.Airport)]
+	//public class AirportNew(Dictionary<LanguageId, List<string>> name, AirportObject airport, List<G1Element32> graphics)
+	//{
+	//	public Dictionary<LanguageId, List<string>> Name { get; } = name ?? throw new ArgumentNullException(nameof(name));
+
+	//	public int16_t BuildCostFactor { get; set; } = airport.BuildCostFactor;
+	//	public int16_t SellCostFactor { get; set; } = airport.SellCostFactor;
+	//	public uint8_t CostIndex { get; set; } = airport.CostIndex;
+	//	public uint16_t AllowedPlaneTypes { get; set; } = airport.AllowedPlaneTypes;
+	//	public uint8_t NumBuildingAnimations { get; set; } = airport.NumBuildingAnimations;
+	//	public int16_t NumBuildingVariations { get; set; } = airport.NumBuildingVariations;
+	//	public IList<uint8_t> BuildingVariationHeights { get; } = airport.BuildingVariationHeights ?? throw new ArgumentNullException(nameof(airport.BuildingVariationHeights));
+	//	public IList<BuildingPartAnimation> BuildingVariationAnimations { get; } = airport.BuildingVariationAnimations ?? throw new ArgumentNullException(nameof(airport.BuildingVariationAnimations));
+	//	public IList<uint8_t[]> BuildingVariationParts { get; } = airport.BuildingVariationParts ?? throw new ArgumentNullException(nameof(airport.BuildingVariationParts));
+	//	public IList<AirportBuilding> BuildingPositions { get; } = airport.BuildingPositions ?? throw new ArgumentNullException(nameof(airport.BuildingPositions));
+	//	public uint32_t LargeTiles { get; set; } = airport.LargeTiles;
+	//	public int8_t MinX { get; set; } = airport.MinX;
+	//	public int8_t MinY { get; set; } = airport.MinY;
+	//	public int8_t MaxX { get; set; } = airport.MaxX;
+	//	public int8_t MaxY { get; set; } = airport.MaxY;
+	//	public uint16_t DesignedYear { get; set; } = airport.DesignedYear;
+	//	public uint16_t ObsoleteYear { get; set; } = airport.ObsoleteYear;
+	//	public uint8_t NumMovementNodes { get; set; } = airport.NumMovementNodes;
+	//	public uint8_t NumMovementEdges { get; set; } = airport.NumMovementEdges;
+	//	public IList<MovementNode> MovementNodes { get; } = airport.MovementNodes ?? throw new ArgumentNullException(nameof(airport.MovementNodes));
+	//	public IList<MovementEdge> MovementEdges { get; } = airport.MovementEdges ?? throw new ArgumentNullException(nameof(airport.MovementEdges));
+
+	//	public IList<G1Element32> Graphics { get; } = graphics ?? throw new ArgumentNullException(nameof(graphics));
+
+	//	//public const int FixedSize = 0xBA;
+
+	//	//public static AirportNew FromDatFile(ReadOnlySpan<byte> bytes) // bytes is the dat file, minus the s5 and obj headers
+	//	//{
+	//	//	var x = new AirportNew();
+	//	//	x.BuildCostFactor = ByteReaderT.Read_int16t(bytes, 0x02);
+	//	//	x.SellCostFactor = ByteReaderT.Read_int16t(bytes, 0x04);
+	//	//	x.CostIndex = ByteReaderT.Read_uint8t(bytes, 0x06);
+	//	//	x.AllowedPlaneTypes = ByteReaderT.Read_uint16t(bytes, 0x10);
+	//	//	x.NumBuildingAnimations = ByteReaderT.Read_uint8t(bytes, 0x12);
+	//	//	x.NumBuildingVariations = ByteReaderT.Read_uint8t(bytes, 0x13);
+
+	//	//	//x.BuildingVariationHeights = ByteReaderT.Read_Array<uint8_t>(bytes, 0x14);
+	//	//	//x.BuildingVariationAnimations = ByteReaderT.Read_int16t(bytes, 0x18);
+	//	//	//x.BuildingVariationParts = ByteReaderT.Read_uint8t(bytes, 0x1C);
+	//	//	//x.BuildingPositions = ByteReaderT.Read_int16t(bytes, 0x9C);
+
+	//	//	x.LargeTiles = ByteReaderT.Read_uint32t(bytes, 0xA0);
+	//	//	x.MinX = ByteReaderT.Read_int8t(bytes, 0xA4);
+	//	//	x.MinY = ByteReaderT.Read_int8t(bytes, 0xA5);
+	//	//	x.MaxX = ByteReaderT.Read_int8t(bytes, 0xA6);
+	//	//	x.MaxY = ByteReaderT.Read_int8t(bytes, 0xA7);
+	//	//	x.DesignedYear = ByteReaderT.Read_uint16t(bytes, 0xA8);
+	//	//	x.ObsoleteYear = ByteReaderT.Read_uint16t(bytes, 0xAA);
+	//	//	x.NumMovementNodes = ByteReaderT.Read_uint8t(bytes, 0xAC);
+	//	//	x.NumMovementEdges = ByteReaderT.Read_uint8t(bytes, 0xAD);
+
+	//	//	///x.MovementNodes = ByteReaderT.Read_int16t(bytes, 0x02);
+	//	//	//x.MovementEdgesMaxY = ByteReaderT.Read_int16t(bytes, 0x04);
+
+	//	//	// move past fixed section
+	//	//	bytes = bytes[FixedSize..];
+
+	//	//	// read variable data
+
+	//	//	// read string table
+	//	//	var bytesString = 1;
+	//	//	bytes = bytes[bytesString..];
+
+	//	//	// read graphics
+
+	//	//	return x;
+	//	//}
+	//}
 
 	[TypeConverter(typeof(ExpandableObjectConverter))]
 	[LocoStructSize(0xBA)]
@@ -107,9 +168,8 @@ namespace Core.Objects
 			MovementEdges.AddRange(ByteReader.ReadLocoStructArray(remainingData[..(NumMovementEdges * edgeSize)], typeof(MovementEdge), NumMovementEdges, edgeSize)
 				.Cast<MovementEdge>()
 				.ToList());
-			remainingData = remainingData[(NumMovementEdges * edgeSize)..];
 
-			return remainingData;
+			return remainingData[(NumMovementEdges * edgeSize)..];
 		}
 
 		public ReadOnlySpan<byte> Save()
