+++ Core/Objects/Sound/SoundObject.cs
@@ -3,7 +3,7 @@ using OpenLoco.Dat.Data;
 using OpenLoco.Dat.FileParsing;
 using OpenLoco.Dat.Types;
 
-namespace Core.Objects.Sound
+namespace OpenLoco.Dat.Objects.Sound
 {
 	[TypeConverter(typeof(ExpandableObjectConverter))]
 	[LocoStructSize(0x0C)]
