+++ Core/Types/SCV5/LandGeneratorType.cs
@@ -1,4 +1,4 @@
-namespace Core.Types.SCV5
+namespace OpenLoco.Dat.Types.SCV5
 {
 	public enum LandGeneratorType : uint8_t
 	{
