+++ Core/Types/SCV5/ObjectiveType.cs
@@ -1,4 +1,4 @@
-namespace Core.Types.SCV5
+namespace OpenLoco.Dat.Types.SCV5
 {
 	public enum ObjectiveType : uint8_t
 	{
