+++ Core/Types/SCV5/S5FixFlags.cs
@@ -1,4 +1,4 @@
-namespace Core.Types.SCV5
+namespace OpenLoco.Dat.Types.SCV5
 {
 	[Flags]
 	public enum S5FixFlags : uint16_t
