+++ Core/Types/SCV5/TopographyStyle.cs
@@ -1,4 +1,4 @@
-namespace Core.Types.SCV5
+namespace OpenLoco.Dat.Types.SCV5
 {
 	public enum TopographyStyle : uint8_t
 	{
