+++ Tests/LoadSaveTests.cs
@@ -1,11 +1,10 @@
-using Core.Objects;
-using Core.Objects.Sound;
+using OpenLoco.Dat.Objects;
+using OpenLoco.Dat.Objects.Sound;
 using NUnit.Framework;
 using NUnit.Framework.Internal;
 using OpenLoco.Dat.Data;
 using OpenLoco.Dat.FileParsing;
 using OpenLoco.Dat.Types;
-using OpenLoco.Dat.Objects;
 using Logger = OpenLoco.Dat.Logging.Logger;
 
 namespace OpenLoco.Dat.Tests
